+++ CHANGES.md
@@ -23,6 +23,7 @@ Changes in 13.0
 - Deprecate HackingLog() function in HackingLog.fnc.php
 - Add RosarioSIS\Functions\Hacking class in Hacking.php
 - Use RosarioSIS\Functions\Hacking class, program wide
+- Remove deprecated `_printPageHead()` function in PasswordReset.php
 
 Changes in 12.9.2
 -----------------